+++ src/ImobFeed.Core/Leitores/LeitorRecomendacaoItau.cs
@@ -21,7 +21,7 @@ public sealed class LeitorRecomendacaoItau : ILeitorRecomendacao
             if (codigo is null)
                 continue;
 
-            float? peso = LeitorCampo.LerPeso(line);
+            decimal? peso = LeitorCampo.LerPeso(line);
             if (peso is null)
                 continue;
